requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-pinecone>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pinecone-client>=5.0.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
